Cleanup
- Remove error scopes
- Replace duplicate declares with fwd.hpp include
