--- src/gfx/async.hpp
@@ -1,11 +1,0 @@
-#pragma once
-
-namespace gfx {
-
-template <typename T> struct ContextData {};
-
-struct AsyncMap {
-
-};
-
-} // namespace gfx
