+++ src/gfx/feature.hpp
@@ -2,6 +2,7 @@
 #include "enums.hpp"
 #include "linalg.hpp"
 #include "params.hpp"
+#include "fwd.hpp"
 #include <functional>
 #include <memory>
 #include <optional>
@@ -136,11 +137,6 @@ struct NamedShaderParam {
 		hasher(flags);
 	}
 };
-
-struct Drawable;
-struct View;
-struct Context;
-typedef std::shared_ptr<View> ViewPtr;
 
 struct FeatureCallbackContext {
 	Context &context;
