+++ src/gfx/mesh.hpp
@@ -85,6 +85,5 @@ protected:
 	void initContextData(Context &context, MeshContextData &contextData);
 	void releaseContextData();
 };
-using MeshPtr = std::shared_ptr<Mesh>;
 
 } // namespace gfx
