--- src/gfx/model.hpp
@@ -1,15 +1,0 @@
-#pragma once
-
-#include "chainblocks.hpp"
-#include "mesh.hpp"
-#include <linalg/linalg.h>
-#include <linalg_shim.hpp>
-#include <vector>
-
-namespace tinygltf {
-struct Animation;
-}
-
-namespace gfx {
-// struct Mesh {};
-} // namespace gfx
